Fix several bugs and add workarounds
Now the repositories are cloned to the staff namespace directly,
which prevents the issue of running out of personal project space
(and is much tidier for the one doing the cloning).

Also the default branch is now unprotected to allow developers to
push to the cloned repositories, plus a few smaller bug fixes.